requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
async-timeout==4.0.3
decorator==5.2.1
exceptiongroup==1.2.2
executing==2.2.0
filelock==3.13.1
fsspec==2024.6.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.3
imageio==2.37.0
ipdb==0.13.13
ipython==8.34.0
jedi==0.19.2
jinja2==3.1.4
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.21
langchain-core==0.3.47
langchain-text-splitters==0.3.7
langsmith==0.3.18
lazy-loader==0.4
matplotlib-inline==0.1.7
mpmath==1.3.0
networkx==3.3
nvidia-cublas-cu12==12.4.2.65
nvidia-cuda-cupti-cu12==12.4.99
nvidia-cuda-nvrtc-cu12==12.4.99
nvidia-cuda-runtime-cu12==12.4.99
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.0.44
nvidia-curand-cu12==10.3.5.119
nvidia-cusolver-cu12==11.6.0.99
nvidia-cusparse-cu12==12.3.0.142
nvidia-ml-py==12.570.86
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.4.99
nvitop==1.4.2
orjson==3.10.15
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
prompt-toolkit==3.0.50
psutil==7.0.0
ptyprocess==0.7.0
pure-eval==0.2.3
pydantic==2.10.6
pydantic-core==2.27.2
pygments==2.19.1
python-dateutil==2.9.0.post0
pytz==2025.1
pyyaml==6.0.2
requests-toolbelt==1.0.0
safetensors==0.5.3
scikit-image==0.25.2
sniffio==1.3.1
sqlalchemy==2.0.39
stack-data==0.6.3
sympy==1.13.1
tenacity==9.0.0
tifffile==2025.3.13
tomli==2.2.1
torchxrayvision==1.3.3
tqdm==4.67.1
traitlets==5.14.3
triton==3.0.0
tzdata==2025.2
wcwidth==0.2.13
zstandard==0.23.0